access setting

+++ application/views/client/client_list.php
@@ -46,15 +46,21 @@
             </header>
             <div class="panel-tools fully" align="right" data-toolscolor="#6CC3A0">
                 <ul class="tooltip-area">
+                    <?php if($this->template->checkRole($this->session->userdata('userlevelId'),'client','create')){ ?>  
                     <li>
                         <a href="<?= base_url('client/create'); ?>" class="btn btn-success" title="create new data"><i class="fa fa-plus-square"></i></a>
                     </li>
+                    <?php } ?>
+                    <?php if($this->template->checkRole($this->session->userdata('userlevelId'),'client','excel')){ ?>
                     <li>
                         <a href="<?= base_url('client/excel'); ?>" class="btn btn-theme-inverse" title="download excel"><i class="fa fa-print"></i></a>
                     </li>
+                     <?php } ?>
+                    <?php if($this->template->checkRole($this->session->userdata('userlevelId'),'client','word')){ ?>
                     <li>
                         <a href="<?= base_url('client/word'); ?>" class="btn btn-warning" title="download word"><i class="fa fa-file-text"></i></a>
                     </li>
+                     <?php } ?>
                     <li></li>
                     <li><a href="javascript:void(0)" class="btn btn-collapse" title="Collapse"><i class="fa fa-sort-amount-asc"></i></a></li>
                     <li><a href="javascript:void(0)" class="btn btn-reload"  title="Reload"><i class="fa fa-retweet"></i></a></li>
@@ -97,15 +103,21 @@
                         <td><?php echo $clientStatusBtn ?></td>
 			<td style="text-align:center" width="200px">
                             <span class="tooltip-area">
+                                  <?php if($this->template->checkRole($this->session->userdata('userlevelId'),'client','update')){ ?>  
                                     <a href="<?= base_url() ?>client/update/<?=$client->id;?>" class="btn btn-default btn-sm" title="Edit">
                                         <i class="fa fa-pencil"></i>
                                     </a>
+                                     <?php } ?>
+                                <?php if($this->template->checkRole($this->session->userdata('userlevelId'),'client','view')){ ?>  
                                     <a href="<?= base_url() ?>client/read/<?=$client->id;?>" class="btn btn-default btn-sm" title="detail">
                                         <i class="fa fa-eye"></i>
                                     </a>
+                                     <?php } ?>
+                                      <?php if($this->template->checkRole($this->session->userdata('userlevelId'),'client','delete')){ ?>  
                                     <a href="<?= base_url() ?>client/delete/<?=$client->id;?>"  class="btn btn-default btn-sm" title="Delete">
                                         <i class="fa fa-trash-o"></i>
                                     </a>
+                                     <?php } ?>
                                 </span>
 			</td>
 		</tr>
